f7

+++ src/db/actions.ts
@@ -1,15 +1,28 @@
-'use server'
+"use server";
 
 import { db } from "@/db";
 import { snippets, users } from "@/db/schema";
 import { eq } from "drizzle-orm";
 import { nanoid } from "nanoid";
-import {  CreateUser,  UpdateUser } from "./types";
-import {  createSnippetSchema,  updateSnippetSchema } from "@/lib/validations/snippets";
+import {
+  createSnippetSchema,
+  updateSnippetSchema,
+} from "@/lib/validations/snippets";
 import { redirect } from "next/navigation";
 import { revalidatePath } from "next/cache";
+import { hash, compare } from "bcryptjs";
+import {
+  loginSchema,
+  signupSchema,
+  verifyEmailSchema,
+} from "@/lib/validations/auth";
+import { sendVerificationEmail } from "@/lib/email"; // You'll need to implement this
+
+import { createSession, deleteSession, verifySession } from "@/lib/session";
+import { updateUserSchema } from "@/lib/validations/user";
 
 export type FormState = {
+  success?: boolean;
   errors?: {
     message?: string;
     [key: string]: string[] | string | undefined;
@@ -21,35 +34,14 @@ async function wait(ms: number) {
 }
 
 // User Actions
-export async function createUser(data: CreateUser) {
-  const userId = nanoid();
-  await db.insert(users).values({
-    id: userId,
-    ...data,
-  });
-  return userId;
-}
-
-export async function updateUser({ id, ...data }: UpdateUser) {
-  await db.update(users)
-    .set({
-      ...data,
-      updatedAt: new Date(),
-    })
-    .where(eq(users.id, id));
-  return id;
-}
-
-export async function deleteUser(id: string) {
-  await db.delete(users).where(eq(users.id, id));
-}
-
-// Snippet Actions
-export async function createSnippet(prevState: FormState, formData: FormData): Promise<FormState> {
-  await wait(2000)
-  const formDataObj = Object.fromEntries(formData.entries())
-  const parsedFormData = createSnippetSchema.safeParse(formDataObj)
-  console.log(parsedFormData.data?.userId);
+export async function updateUser(
+  prevState: FormState,
+  formData: FormData
+): Promise<FormState> {
+  await wait(2000);
+  const { userId } = await verifySession();
+  const formDataObj = Object.fromEntries(formData.entries());
+  const parsedFormData = updateUserSchema.safeParse(formDataObj);
 
   if (!parsedFormData.success) {
     return {
@@ -57,14 +49,27 @@ export async function createSnippet(prevState: FormState, formData: FormData): P
     };
   }
 
-  const snippetId = nanoid();
-
   try {
-    await db.insert(snippets).values({
-      id: snippetId,
-      ...parsedFormData.data,
+    const existingUser = await db.query.users.findFirst({
+      where: eq(users.username, parsedFormData.data.username),
     });
-  } catch  {
+
+    if (existingUser) {
+      return {
+        errors: {
+          username: "این نام کاربری قبلاً استفاده شده است",
+        },
+      };
+    }
+
+    await db
+      .update(users)
+      .set({
+        ...parsedFormData.data,
+        updatedAt: new Date(),
+      })
+      .where(eq(users.id, userId));
+  } catch {
     return {
       errors: {
         message: "خطایی رخ داده است",
@@ -72,17 +77,19 @@ export async function createSnippet(prevState: FormState, formData: FormData): P
     };
   }
 
-  revalidatePath('/snippets')
-  revalidatePath(`/users/${parsedFormData.data.userId}`)
-  revalidatePath(`/snippets/${snippetId}`)
-  redirect(`/snippets/${snippetId}`)
+  revalidatePath(`/dashboard`);
+  revalidatePath(`/users`);
+  revalidatePath(`/users/${userId}`);
+  redirect(`/dashboard`);
 }
 
-
-export async function updateSnippet(prevState: FormState, formData: FormData): Promise<FormState> {
-  await wait(2000)
-  const formDataObj = Object.fromEntries(formData.entries())
-  const parsedFormData = updateSnippetSchema.safeParse(formDataObj)
+// Snippet Actions
+export async function createSnippet(
+  prevState: FormState,
+  formData: FormData
+): Promise<FormState> {
+  const formDataObj = Object.fromEntries(formData.entries());
+  const parsedFormData = createSnippetSchema.safeParse(formDataObj);
 
   if (!parsedFormData.success) {
     return {
@@ -90,13 +97,49 @@ export async function updateSnippet(prevState: FormState, formData: FormData): P
     };
   }
 
-  
+  const { userId } = await verifySession();
+  const snippetId = nanoid();
+
+  try {
+    await db.insert(snippets).values({
+      id: snippetId,
+      userId,
+      ...parsedFormData.data,
+    });
+  } catch {
+    return {
+      errors: {
+        message: "خطایی رخ داده است",
+      },
+    };
+  }
+
+  revalidatePath("/snippets");
+  revalidatePath(`/users/${userId}`);
+  redirect(`/snippets/${snippetId}`);
+}
+
+export async function updateSnippet(
+  prevState: FormState,
+  formData: FormData
+): Promise<FormState> {
+  await wait(2000);
+  const formDataObj = Object.fromEntries(formData.entries());
+  const parsedFormData = updateSnippetSchema.safeParse(formDataObj);
+
+  if (!parsedFormData.success) {
+    return {
+      errors: parsedFormData.error.flatten().fieldErrors,
+    };
+  }
+
+  const { userId } = await verifySession();
 
   const selectedSnippet = await db.query.snippets.findFirst({
-    where: eq(snippets.id, parsedFormData.data.id)
-  })
+    where: eq(snippets.id, parsedFormData.data.id),
+  });
 
- const isOwner = selectedSnippet?.userId === parsedFormData.data.userId
+  const isOwner = selectedSnippet?.userId === userId;
 
   if (!isOwner) {
     return {
@@ -107,11 +150,14 @@ export async function updateSnippet(prevState: FormState, formData: FormData): P
   }
 
   try {
-    await db.update(snippets).set({
-      ...parsedFormData.data,
-      updatedAt: new Date(),
-    }).where(eq(snippets.id, parsedFormData.data.id))
-  } catch  {
+    await db
+      .update(snippets)
+      .set({
+        ...parsedFormData.data,
+        updatedAt: new Date(),
+      })
+      .where(eq(snippets.id, parsedFormData.data.id));
+  } catch {
     return {
       errors: {
         message: "خطایی رخ داده است",
@@ -119,19 +165,16 @@ export async function updateSnippet(prevState: FormState, formData: FormData): P
     };
   }
 
-  revalidatePath('/snippets')
-  revalidatePath(`/snippets/${parsedFormData.data.id}`)
-  revalidatePath(`/users/${parsedFormData.data.userId}`)
-  redirect(`/snippets/${parsedFormData.data.id}`)
+  revalidatePath("/snippets");
+  revalidatePath(`/snippets/${parsedFormData.data.id}`);
+  revalidatePath(`/users/${userId}`);
+  redirect(`/snippets/${parsedFormData.data.id}`);
 }
 
 export async function deleteSnippet(id: string): Promise<FormState> {
-  
   const selectedSnippet = await db.query.snippets.findFirst({
-    where: eq(snippets.id, id)
-  })
-
-  console.log(selectedSnippet)
+    where: eq(snippets.id, id),
+  });
 
   if (!selectedSnippet) {
     return {
@@ -142,6 +185,16 @@ export async function deleteSnippet(id: string): Promise<FormState> {
   }
 
   try {
+    const { userId } = await verifySession();
+
+    if (selectedSnippet.userId !== userId) {
+      return {
+        errors: {
+          message: "شما نمیتوانید قطعه کد دیگری را حذف کنید",
+        },
+      };
+    }
+
     await db.delete(snippets).where(eq(snippets.id, id));
   } catch {
     return {
@@ -151,7 +204,252 @@ export async function deleteSnippet(id: string): Promise<FormState> {
     };
   }
 
-  revalidatePath('/snippets')
-  revalidatePath(`/users/${selectedSnippet.userId}`)
-  redirect(`/users/${selectedSnippet.userId}`)
-} 
+  revalidatePath("/snippets");
+  revalidatePath(`/users/${selectedSnippet.userId}`);
+  redirect(`/users/${selectedSnippet.userId}`);
+}
+
+// Auth Actions
+export async function signup(
+  prevState: FormState,
+  formData: FormData
+): Promise<FormState> {
+  const formDataObj = Object.fromEntries(formData.entries());
+  const parsedData = signupSchema.safeParse(formDataObj);
+
+  if (!parsedData.success) {
+    return {
+      errors: parsedData.error.flatten().fieldErrors,
+    };
+  }
+
+  try {
+    // Generate verification code
+    const verificationCode = Math.floor(
+      100000 + Math.random() * 900000
+    ).toString();
+    const verificationCodeExpiresAt = new Date(Date.now() + 10 * 60 * 1000); // 10 minutes
+
+    const hashedPassword = await hash(parsedData.data.password, 10);
+
+    // Check if user exists
+    const existingUser = await db.query.users.findFirst({
+      where: eq(users.email, parsedData.data.email),
+    });
+
+    if (existingUser && existingUser.emailVerified) {
+      return {
+        errors: {
+          email: "این ایمیل قبلاً ثبت شده است",
+        },
+      };
+    } else if (existingUser && !existingUser.emailVerified) {
+      await db
+        .update(users)
+        .set({
+          name: parsedData.data.name,
+          verificationCode,
+          verificationCodeExpiresAt,
+          password: hashedPassword,
+        })
+        .where(eq(users.id, existingUser.id));
+    } else {
+      // Create user
+      const userId = nanoid();
+      const username = `user-${Math.floor(100000 + Math.random() * 900000)}`;
+      const data = await db
+        .insert(users)
+        .values({
+          id: userId,
+          name: parsedData.data.name,
+          username,
+          email: parsedData.data.email,
+          password: hashedPassword,
+          verificationCode,
+          verificationCodeExpiresAt,
+          deleteAt: new Date(Date.now() + 24 * 60 * 60 * 1000), // Delete after 24 hours if not verified
+        })
+        .returning({ id: users.id });
+
+      const user = data[0];
+
+      if (!user) {
+        return {
+          errors: {
+            message: "خطایی رخ داده است",
+          },
+        };
+      }
+    }
+
+    // Send verification email
+    await sendVerificationEmail(parsedData.data.email, verificationCode);
+  } catch {
+    return {
+      errors: {
+        message: "خطایی رخ داده است",
+      },
+    };
+  }
+
+  // Redirect to verify page
+  redirect(`/verify-email?email=${parsedData.data.email}`);
+}
+
+export async function login(
+  prevState: FormState,
+  formData: FormData
+): Promise<FormState> {
+  const formDataObj = Object.fromEntries(formData.entries());
+  const parsedData = loginSchema.safeParse(formDataObj);
+
+  if (!parsedData.success) {
+    return {
+      errors: parsedData.error.flatten().fieldErrors,
+    };
+  }
+
+  const user = await db.query.users.findFirst({
+    where: eq(users.email, parsedData.data.email),
+  });
+
+  if (!user) {
+    return {
+      errors: {
+        message: "ایمیل یا رمز عبور اشتباه است",
+      },
+    };
+  }
+
+  const isPasswordCorrect = await compare(
+    parsedData.data.password,
+    user.password
+  );
+
+  if (!isPasswordCorrect) {
+    return {
+      errors: {
+        message: "ایمیل یا رمز عبور اشتباه است",
+      },
+    };
+  }
+
+  await createSession(user.id);
+
+  // Get callback URL from form data
+  const callbackUrl = formData.get("callbackUrl") as string;
+  redirect(callbackUrl || "/dashboard");
+}
+
+export async function verifyEmail(
+  prevState: FormState,
+  formData: FormData
+): Promise<FormState> {
+  const email = formData.get("email") as string;
+  const code = formData.get("code") as string;
+
+  const parsedData = verifyEmailSchema.safeParse({ code });
+  if (!parsedData.success) {
+    return {
+      errors: parsedData.error.flatten().fieldErrors,
+    };
+  }
+
+  let user;
+
+  try {
+    user = await db.query.users.findFirst({
+      where: eq(users.email, email),
+    });
+
+    if (!user || user.verificationCode !== code) {
+      return {
+        errors: {
+          code: "کد تأیید نامعتبر است",
+        },
+      };
+    }
+
+    if (
+      user.verificationCodeExpiresAt &&
+      user.verificationCodeExpiresAt < new Date()
+    ) {
+      return {
+        errors: {
+          code: "کد تأیید منقضی شده است",
+        },
+      };
+    }
+
+    // Verify user
+    await db
+      .update(users)
+      .set({
+        emailVerified: true,
+        verificationCode: null,
+        verificationCodeExpiresAt: null,
+      })
+      .where(eq(users.id, user.id));
+
+    // Create session token
+    await createSession(user.id);
+  } catch {
+    return {
+      errors: {
+        message: "خطایی رخ داده است",
+      },
+    };
+  }
+
+  redirect(`/dashboard`);
+}
+
+export async function resendVerificationCode(
+  prevState: FormState,
+  formData: FormData
+): Promise<FormState> {
+  const email = formData.get("email") as string;
+
+  try {
+    const user = await db.query.users.findFirst({
+      where: eq(users.email, email),
+    });
+
+    if (!user || user.emailVerified) {
+      return {
+        errors: {
+          message: "کاربر یافت نشد یا قبلاً تأیید شده است",
+        },
+      };
+    }
+
+    const verificationCode = Math.floor(
+      100000 + Math.random() * 900000
+    ).toString();
+    const verificationCodeExpiresAt = new Date(Date.now() + 10 * 60 * 1000);
+
+    await db
+      .update(users)
+      .set({
+        verificationCode,
+        verificationCodeExpiresAt,
+        deleteAt: new Date(Date.now() + 24 * 60 * 60 * 1000), // Extend deletion time
+      })
+      .where(eq(users.id, user.id));
+
+    await sendVerificationEmail(email, verificationCode);
+
+    return {};
+  } catch {
+    return {
+      errors: {
+        message: "خطایی رخ داده است",
+      },
+    };
+  }
+}
+
+export async function logout() {
+  deleteSession();
+  redirect("/login");
+}
